+++ Engine/Src/Ancona/Core2D/Systems/Drawable/ContainerDrawable.hpp
@@ -30,9 +30,9 @@ class ContainerDrawable : public Drawable
          *
          */
         ContainerDrawable(
-                const int priority,
+                const float priority,
                 const std::string & key,
-                int priorityOffset = 0,
+                float priorityOffset = 0,
                 sf::Vector2f anchor = sf::Vector2f(0.0f, 0.0f));
 
         Drawable * Copy() override;
